The non-data size of Twrite messages is 23 bytes (not 24)
The manpage describes Twrite as:

size[4] Twrite tag[2] fid[4] offset[8] count[4] data[count]

if count = 0, the message size is 4 + 1 + 2 + 4 + 8 + 4 = 23

This manifests as a bug if a client agrees a maximum message size
with a server, and then computes the maximum value for `count`
by subtracting 23 i.e. the biggest possible write message. If the
server then checks whether the request is too big by subtracting
24, it will return an error.

Signed-off-by: David Scott <dave.scott@unikernel.com>